+++ extensions/data-plane/data-plane-http/src/test/java/org/eclipse/dataspaceconnector/dataplane/http/pipeline/HttpDataSourceFactoryTest.java
@@ -17,6 +17,8 @@
 
 package org.eclipse.dataspaceconnector.dataplane.http.pipeline;
 
+import com.github.javafaker.Faker;
+import io.netty.handler.codec.http.HttpMethod;
 import net.jodah.failsafe.RetryPolicy;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
@@ -26,6 +28,7 @@ import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
 import org.eclipse.dataspaceconnector.spi.types.domain.DataAddress;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.DataFlowRequest;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -47,7 +50,7 @@ import static org.eclipse.dataspaceconnector.dataplane.spi.schema.DataFlowReques
 import static org.eclipse.dataspaceconnector.dataplane.spi.schema.DataFlowRequestSchema.QUERY_PARAMS;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_CODE;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_KEY;
-import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.ENDPOINT;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.BASE_URL;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.NAME;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.PROXY_BODY;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.PROXY_METHOD;
@@ -61,20 +64,29 @@ import static org.mockito.Mockito.when;
 
 class HttpDataSourceFactoryTest {
 
-    private static final String TEST_SECRET_NAME = "secret-test";
-    private static final String TEST_SECRET_VALUE = "token-test";
-
-    private static final OkHttpClient HTTP_CLIENT = mock(OkHttpClient.class);
-    private static final RetryPolicy<Object> RETRY_POLICY = new RetryPolicy<>();
-    private static final Monitor MONITOR = mock(Monitor.class);
-
+    private static final Faker FAKER = new Faker();
+    private static OkHttpClient httpClient;
+    private static RetryPolicy<Object> retryPolicy;
+    private static Monitor monitor;
+    private static String secretName;
+    private static String secretValue;
     private Vault vaultMock;
+
     private HttpDataSourceFactory factory;
+
+    @BeforeAll
+    public static void init() {
+        httpClient = mock(OkHttpClient.class);
+        retryPolicy = new RetryPolicy<>();
+        monitor = mock(Monitor.class);
+        secretName = FAKER.lorem().word();
+        secretValue = FAKER.internet().uuid();
+    }
 
     @BeforeEach
     void setUp() {
         vaultMock = mock(Vault.class);
-        factory = new HttpDataSourceFactory(HTTP_CLIENT, RETRY_POLICY, MONITOR, vaultMock);
+        factory = new HttpDataSourceFactory(httpClient, retryPolicy, monitor, vaultMock);
     }
 
     @ParameterizedTest
@@ -96,7 +108,7 @@ class HttpDataSourceFactoryTest {
     @ParameterizedTest(name = "{index} {0}")
     @MethodSource("provideTestInstances")
     void verifySourceValidationAndCreation(String name, TestInstance testInstance) {
-        when(vaultMock.resolveSecret(TEST_SECRET_NAME)).thenReturn(TEST_SECRET_VALUE);
+        when(vaultMock.resolveSecret(secretName)).thenReturn(secretValue);
 
         var request = testInstance.createRequest();
         var expectedSource = testInstance.createDataSource();
@@ -124,20 +136,20 @@ class HttpDataSourceFactoryTest {
      * Serves some invalid {@link DataFlowRequest}.
      */
     private static Stream<Arguments> provideInvalidRequests() {
-        var endpoint = "http://example.com";
-        var authKey = "apikey-test";
+        var endpoint = FAKER.internet().url();
+        var authKey = FAKER.lorem().word();
 
         var missingEndpoint = TestInstance.newInstance()
-                .method("GET", true);
+                .method(HttpMethod.GET.name(), true);
 
         var incompleteHeader = TestInstance.newInstance()
-                .method("GET", true)
+                .method(HttpMethod.GET.name(), true)
                 .authKey(authKey)
                 .endpoint(endpoint);
 
         var unknownMediaType = TestInstance.newInstance()
-                .method("POST", true)
-                .body("dummy", "hello world!", true)
+                .method(HttpMethod.POST.name(), true)
+                .body(FAKER.lorem().word(), FAKER.internet().uuid(), true)
                 .endpoint(endpoint);
 
         return Stream.of(
@@ -151,71 +163,72 @@ class HttpDataSourceFactoryTest {
      * Serves some valid {@link DataFlowRequest} with the associated expected {@link HttpDataSource} that must be generated.
      */
     private static Stream<Arguments> provideTestInstances() {
-        var endpoint = "http://example.com";
-        var name = "foo.json";
-        var authKey = "apikey-test";
+        var endpoint = FAKER.internet().url();
+        var name = FAKER.lorem().word();
+        var path = FAKER.lorem().word();
+        var authKey = FAKER.lorem().word();
         var mediaType = "application/json";
-        var body = "test";
-        var queryParams = "?foo=bar";
+        var body = FAKER.lorem().sentence();
+        var queryParams = FAKER.lorem().word();
 
         var get = TestInstance.newInstance()
-                .method("GET", true)
+                .method(HttpMethod.GET.name(), true)
                 .endpoint(endpoint);
 
         var getWithPath = TestInstance.newInstance()
-                .method("GET", true)
-                .basePath("hello/world", true)
+                .method(HttpMethod.GET.name(), true)
+                .basePath(path, true)
                 .endpoint(endpoint);
 
         var ignorePathIfProxyDisabled = TestInstance.newInstance()
-                .method("GET", true)
-                .name("some/name")
-                .basePath("hello/world", false)
+                .method(HttpMethod.GET.name(), true)
+                .name(name)
+                .basePath(path, false)
                 .endpoint(endpoint);
 
         var getWithName = TestInstance.newInstance()
-                .method("GET", true)
+                .method(HttpMethod.GET.name(), true)
                 .name(name)
                 .endpoint(endpoint);
 
         var getWithQueryParams = TestInstance.newInstance()
-                .method("GET", true)
+                .method(HttpMethod.GET.name(), true)
                 .queryParams(queryParams, true)
                 .endpoint(endpoint);
 
         var ignoreQueryParamsIfProxyDisabled = TestInstance.newInstance()
-                .method("GET", true)
+                .method(HttpMethod.GET.name(), true)
                 .queryParams(queryParams, false)
                 .endpoint(endpoint);
 
         var getWithSecret = TestInstance.newInstance()
-                .method("GET", true)
-                .authHeader(authKey, TEST_SECRET_NAME, TEST_SECRET_VALUE)
+                .method(HttpMethod.GET.name(), true)
+                .authHeader(authKey, secretName, secretValue)
                 .endpoint(endpoint);
 
         var getWithAuthCode = TestInstance.newInstance()
-                .method("GET", true)
-                .authHeader(authKey, TEST_SECRET_VALUE)
+                .method(HttpMethod.GET.name(), true)
+                .authHeader(authKey, secretValue)
                 .endpoint(endpoint);
 
         var post = TestInstance.newInstance()
-                .method("POST", true)
+                .method(HttpMethod.POST.name(), true)
                 .body(mediaType, body, true)
                 .endpoint(endpoint);
 
         var ignoreMethodIfProxyDisabled = TestInstance.newInstance()
-                .method("POST", false)
+                .method(HttpMethod.POST.name(), false)
                 .body(mediaType, body, true)
                 .endpoint(endpoint);
 
 
         var ignoreBodyWithoutMediaType = TestInstance.newInstance()
-                .method("POST", true)
+                .method(HttpMethod.POST.name(), true)
                 .body(body, true)
                 .endpoint(endpoint);
 
         var ignoreBodyIfProxyDisabled = TestInstance.newInstance()
-                .method("POST", true)
+                .method(HttpMethod.POST.name(), true)
                 .body(mediaType, body, false)
                 .endpoint(endpoint);
 
@@ -241,7 +254,7 @@ class HttpDataSourceFactoryTest {
     private static final class TestInstance {
 
         private static final DataAddress DUMMY_ADDRESS = DataAddress.Builder.newInstance()
-                .type("dummy")
+                .type(FAKER.lorem().word())
                 .build();
 
         private final Map<String, String> props;
@@ -270,7 +283,7 @@ class HttpDataSourceFactoryTest {
         }
 
         public TestInstance endpoint(String endpoint) {
-            this.address.property(ENDPOINT, endpoint);
+            this.address.property(BASE_URL, endpoint);
             source.sourceUrl(endpoint);
             return this;
         }
@@ -278,7 +291,7 @@ class HttpDataSourceFactoryTest {
         public TestInstance method(String method, boolean proxy) {
             props.put(METHOD, method);
             address.property(PROXY_METHOD, Boolean.toString(proxy));
-            source.method(proxy ? method : "GET");
+            source.method(proxy ? method : HttpMethod.GET.name());
             return this;
         }
 
@@ -329,10 +342,10 @@ class HttpDataSourceFactoryTest {
         }
 
         public TestInstance basePath(String basePath, boolean proxy) {
-            props.put(PATH, "hello/world");
+            props.put(PATH, basePath);
             address.property(PROXY_PATH, Boolean.toString(proxy));
             if (proxy) {
-                source.name(basePath);
+                source.path(basePath);
             }
             return this;
         }
@@ -346,9 +359,9 @@ class HttpDataSourceFactoryTest {
 
         private HttpDataSource.Builder defaultHttpSource() {
             return HttpDataSource.Builder.newInstance()
-                    .httpClient(HTTP_CLIENT)
-                    .monitor(MONITOR)
-                    .retryPolicy(RETRY_POLICY);
+                    .httpClient(httpClient)
+                    .monitor(monitor)
+                    .retryPolicy(retryPolicy);
         }
     }
 }