+++ extensions/data-plane/integration-tests/src/test/java/org/eclipse/dataspaceconnector/dataplane/http/DataPlaneHttpToHttpIntegrationTests.java
@@ -58,8 +58,8 @@ import static org.eclipse.dataspaceconnector.dataplane.http.HttpTestFixtures.cre
 import static org.eclipse.dataspaceconnector.junit.testfixtures.TestUtils.getFreePort;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_CODE;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_KEY;
-import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.ENDPOINT;
-import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.NAME;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.BASE_URL;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.PROXY_PATH;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.PROXY_QUERY_PARAMS;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.TYPE;
 import static org.hamcrest.Matchers.containsString;
@@ -78,7 +78,7 @@ import static org.mockserver.stop.Stop.stopQuietly;
  * System Test for Data Plane HTTP extension.
  */
 @ComponentTest
-public class DataPlaneHttpIntegrationTests {
+public class DataPlaneHttpToHttpIntegrationTests {
 
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
     private static final Faker FAKER = new Faker();
@@ -96,7 +96,9 @@ public class DataPlaneHttpIntegrationTests {
     private static final String PROCESS_ID = "processId";
     private static final String EDC_TYPE = "edctype";
     private static final String DATA_FLOW_REQUEST_EDC_TYPE = "dataspaceconnector:dataflowrequest";
-    private static final String DPF_HTTP_API_PART_NAME = "sample";
+
+    private static final String DPF_HTTP_API_PATH = "sample";
+    private static final String PROXY_PATH_TOGGLE = "true";
     private static final String PROXY_QUERY_PARAMS_TOGGLE = "true";
     private static final String AUTH_HEADER_KEY = HttpHeaderNames.AUTHORIZATION.toString();
     private static final String SOURCE_AUTH_VALUE = FAKER.lorem().word();
@@ -352,14 +354,14 @@ public class DataPlaneHttpIntegrationTests {
     /**
      * Request payload with query params to initiate DPF transfer.
      *
-     * @param processId ProcessID of transfer.See {@link DataFlowRequest}
+     * @param processId   ProcessID of transfer.See {@link DataFlowRequest}
      * @param queryParams Query params name and value as key-value entries.
      * @return JSON object. see {@link ObjectNode}.
      */
     private ObjectNode transferRequestPayload(String processId, Map<String, String> queryParams) {
-
         var requestProperties = new HashMap<String, String>();
         requestProperties.put(DataFlowRequestSchema.METHOD, HttpMethod.GET.name());
+        requestProperties.put(DataFlowRequestSchema.PATH, DPF_HTTP_API_PATH);
 
         if (!queryParams.isEmpty()) {
             requestProperties.put(DataFlowRequestSchema.QUERY_PARAMS, queryParams.entrySet()
@@ -371,8 +373,8 @@ public class DataPlaneHttpIntegrationTests {
         var sourceDataAddress = createDataAddress(
                 TYPE,
                 Map.of(
-                        ENDPOINT, DPF_HTTP_SOURCE_API_HOST,
-                        NAME, DPF_HTTP_API_PART_NAME,
+                        BASE_URL, DPF_HTTP_SOURCE_API_HOST,
+                        PROXY_PATH, PROXY_PATH_TOGGLE,
                         PROXY_QUERY_PARAMS, PROXY_QUERY_PARAMS_TOGGLE,
                         AUTHENTICATION_KEY, AUTH_HEADER_KEY,
                         AUTHENTICATION_CODE, SOURCE_AUTH_VALUE
@@ -381,7 +383,7 @@ public class DataPlaneHttpIntegrationTests {
         var destinationDataAddress = createDataAddress(
                 TYPE,
                 Map.of(
-                        ENDPOINT, DPF_HTTP_SINK_API_HOST,
+                        BASE_URL, DPF_HTTP_SINK_API_HOST,
                         AUTHENTICATION_KEY, AUTH_HEADER_KEY,
                         AUTHENTICATION_CODE, SINK_AUTH_VALUE
                 )).build();
@@ -466,7 +468,7 @@ public class DataPlaneHttpIntegrationTests {
         return request
                 .withMethod(HttpMethod.GET.name())
                 .withHeader(AUTH_HEADER_KEY, SOURCE_AUTH_VALUE)
-                .withPath("/" + DPF_HTTP_API_PART_NAME);
+                .withPath("/" + DPF_HTTP_API_PATH);
     }
 
     /**
@@ -478,7 +480,6 @@ public class DataPlaneHttpIntegrationTests {
     private HttpRequest postRequest(String responseBody) {
         return request()
                 .withMethod(HttpMethod.POST.name())
-                .withPath("/" + DPF_HTTP_API_PART_NAME)
                 .withHeaders(
                         new Header(HttpHeaderNames.CONTENT_TYPE.toString(),
                                 MediaType.APPLICATION_OCTET_STREAM.toString()),
@@ -499,7 +500,7 @@ public class DataPlaneHttpIntegrationTests {
     /**
      * Mock plain text response from source.
      *
-     * @param statusCode Response status code.
+     * @param statusCode   Response status code.
      * @param responseBody Response body.
      * @return see {@link HttpResponse}
      */