+++ system-tests/e2e-transfer-test/backend-service/src/main/java/org/eclipse/dataspaceconnector/test/e2e/BackendServiceHttpProvisionerController.java
@@ -23,6 +23,7 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.types.TypeManager;
+import org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema;
 
 import java.util.Map;
 
@@ -59,9 +60,8 @@ public class BackendServiceHttpProvisionerController {
                 // this is the data address of the content served by this backend service
                 "contentDataAddress", Map.of(
                         "properties", Map.of(
-                                "type", "HttpData",
-                                "name", "data",
-                                "endpoint", format("http://localhost:%d/api/service", exposedHttpPort)
+                                "type", HttpDataAddressSchema.TYPE,
+                                HttpDataAddressSchema.BASE_URL, format("http://localhost:%d/api/provider/data", exposedHttpPort)
                         )
                 ),
                 "apiKeyJwt", "unused",