+++ system-tests/e2e-transfer-test/backend-service/src/main/java/org/eclipse/dataspaceconnector/test/e2e/ProviderBackendApiController.java
@@ -1,0 +1,33 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.test.e2e;
+
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+
+import java.util.Map;
+
+@Path("/provider")
+public class ProviderBackendApiController {
+    
+    @Path("/data")
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    public Map<String, String> getData() {
+        return Map.of("message", "some information");
+    }
+}