+++ system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/dataspaceconnector/test/e2e/AbstractEndToEndTransfer.java
@@ -22,8 +22,8 @@ import java.time.Duration;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
-import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.ENDPOINT;
-import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.NAME;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.BASE_URL;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.TYPE;
 import static org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcessStates.COMPLETED;
 import static org.hamcrest.CoreMatchers.equalTo;
 
@@ -60,7 +60,7 @@ public abstract class AbstractEndToEndTransfer {
             given()
                     .baseUri(CONSUMER.backendService().toString())
                     .when()
-                    .get("/api/service/providerData")
+                    .get("/api/consumer/data")
                     .then()
                     .statusCode(200)
                     .body("message", equalTo("some information"));
@@ -96,7 +96,7 @@ public abstract class AbstractEndToEndTransfer {
             given()
                     .baseUri(CONSUMER.backendService().toString())
                     .when()
-                    .get("/api/service/providerData")
+                    .get("/api/consumer/data")
                     .then()
                     .statusCode(200)
                     .body("message", equalTo("some information"));
@@ -120,9 +120,8 @@ public abstract class AbstractEndToEndTransfer {
         assertThat(contractAgreementId).isNotEmpty();
 
         var destination = DataAddress.Builder.newInstance()
-                .type("HttpData")
-                .property(NAME, "data")
-                .property(ENDPOINT, CONSUMER.backendService() + "/api/service")
+                .type(TYPE)
+                .property(BASE_URL, CONSUMER.backendService() + "/api/consumer/store")
                 .build();
         var transferProcessId = CONSUMER.dataRequest(contractAgreementId, assetId, PROVIDER, destination);
 
@@ -135,7 +134,7 @@ public abstract class AbstractEndToEndTransfer {
             given()
                     .baseUri(CONSUMER.backendService().toString())
                     .when()
-                    .get("/api/service/providerData")
+                    .get("/api/consumer/data")
                     .then()
                     .statusCode(200)
                     .body("message", equalTo("some information"));